#81 WIP organizing back end chess code for checkmate solution
