+++ src/backend/services/chess/mateEvaluator.py
@@ -1,7 +1,8 @@
-import copy
+from copy import deepcopy as deepCopy
 
 from src.backend.services.chess.chessConsts import royalOffsets, rookOffsets, bishopOffsets
-from src.backend.utils import hasColorPiece, getPiece, getKingCoords, inCheck, outOfBounds, knightOffsets
+from src.backend.services.chess.chessUtils import hasColorPiece, getPiece, getKingCoords, inCheck, outOfBounds, \
+    knightOffsets
 
 
 # determines if a player cannot make any moves.
@@ -15,28 +16,22 @@ def hasNoLegalMoves(boardstate, playerColor):
                     return False # early return if any move at all is found
     return True
 
-def canMove(boardstate, coords, pieceColor):
+def canMove(boardstate, coords, pieceColor) -> bool:
     enemyColor = "Black" if pieceColor == "White" else "White"
     piece = getPiece(boardstate, coords)
-    pieceType = piece.get("type")
 
-    if pieceType == "Knight":
-        return knightCanMove(boardstate, coords, pieceColor, enemyColor)
-    elif pieceType == "Pawn":
-        return pawnCanMove(boardstate, coords, pieceColor, enemyColor)
-    elif pieceType == "Rook":
-        return rookCanMove(boardstate, coords, pieceColor, enemyColor)  # only need to check adjacent tiles
-    elif pieceType == "Queen":
-        return queenCanMove(boardstate, coords, pieceColor, enemyColor)  # only need to check adjacent tiles
-    elif pieceType == "King":
-        return kingCanMove(boardstate, coords, pieceColor, enemyColor)
-    elif pieceType == "Bishop":
-        return bishopCanMove(boardstate, coords, pieceColor, enemyColor)  # only need to check adjacent tiles
-    return None
+    match piece.get("type"):
+        case "Knight": return knightCanMove(boardstate, coords, pieceColor, enemyColor)
+        case "Pawn":   return pawnCanMove(boardstate, coords, pieceColor, enemyColor)
+        case "Rook":   return rookCanMove(boardstate, coords, pieceColor, enemyColor)  # only need to check adjacent tiles
+        case "Queen":  return queenCanMove(boardstate, coords, pieceColor, enemyColor)  # only need to check adjacent tiles
+        case "King":   return kingCanMove(boardstate, coords, pieceColor, enemyColor)
+        case "Bishop": return bishopCanMove(boardstate, coords, pieceColor, enemyColor)  # only need to check adjacent tiles
+        case _: return False
 
 
 def knightCanMove(boardstate, coords, pieceColor, enemyColor):
-    newBoardstate = copy.deepcopy(boardstate)
+    newBoardstate = deepCopy(boardstate)
     newBoardstate[coords[0]][coords[1]] = {}
     if inCheck(newBoardstate, enemyColor, getKingCoords(newBoardstate, pieceColor)):
         return False
@@ -50,7 +45,7 @@ def knightCanMove(boardstate, coords, pieceColor, enemyColor):
 
 
 def pawnCanMove(boardstate, coords, pieceColor, enemyColor):
-    newBoardstate = copy.deepcopy(boardstate)
+    newBoardstate = deepCopy(boardstate)
     newBoardstate[coords[0]][coords[1]] = {}
     if inCheck(newBoardstate, enemyColor, getKingCoords(newBoardstate, pieceColor)):
         return False
@@ -70,23 +65,28 @@ def pawnCanMove(boardstate, coords, pieceColor, enemyColor):
 
 
 def queenCanMove(boardstate, coords, pieceColor, enemyColor):
-    newBoardstate = copy.deepcopy(boardstate)
-    newBoardstate[coords[0]][coords[1]] = {}
-    if inCheck(newBoardstate, enemyColor, getKingCoords(newBoardstate, pieceColor)):
-        return False
 
-    for offset in royalOffsets:
-        targetCoords = (coords[0] + offset[0], coords[1] + offset[1])
-        if not outOfBounds(targetCoords) and not hasColorPiece(boardstate, targetCoords, pieceColor):
-            return True
     return False
+
+    # TODO code this function according to the following pseudocode
+
+    # validMoveTargets = []
+    # srcCoords = [activePiece.row, activePiece.col];
+    #
+    # offsets.forEach(offset => {
+    #     const rowOffset = offset[0];
+    # const colOffset = offset[1];
+    # scan(boardstate, srcCoords, rowOffset, colOffset, activePiece.row, activePiece.col, activePiece.color, validMoveTargets);
+    # });
+    #
+    # return len(validMoveTargets) > 0;
 
 
 def kingCanMove(boardstate, coords, pieceColor, enemyColor):
     for offset in royalOffsets:
         targetCoords = (coords[0] + offset[0], coords[1] + offset[1])
         if not outOfBounds(targetCoords) and not hasColorPiece(boardstate, targetCoords, pieceColor):
-            newBoardstate = copy.deepcopy(boardstate)
+            newBoardstate = deepCopy(boardstate)
             newBoardstate[coords[0]][coords[1]] = {}
             newBoardstate[targetCoords[0]][targetCoords[1]] = {"piece": {"color": pieceColor, "type": "King"}}
             if not inCheck(newBoardstate, enemyColor, getKingCoords(newBoardstate, pieceColor)):
@@ -95,7 +95,7 @@ def kingCanMove(boardstate, coords, pieceColor, enemyColor):
 
 
 def rookCanMove(boardstate, coords, pieceColor, enemyColor):
-    newBoardstate = copy.deepcopy(boardstate)
+    newBoardstate = deepCopy(boardstate)
     newBoardstate[coords[0]][coords[1]] = {}
     if inCheck(newBoardstate, enemyColor, getKingCoords(newBoardstate, pieceColor)):
         return False
@@ -108,7 +108,7 @@ def rookCanMove(boardstate, coords, pieceColor, enemyColor):
 
 
 def bishopCanMove(boardstate, coords, pieceColor, enemyColor):
-    newBoardstate = copy.deepcopy(boardstate)
+    newBoardstate = deepCopy(boardstate)
     newBoardstate[coords[0]][coords[1]] = {}
     if inCheck(newBoardstate, enemyColor, getKingCoords(newBoardstate, pieceColor)):
         return False
